requirements.txt
# Add your Python dependencies here 
python-dotenv
python-telegram-bot
PyGithub
